pyproject.toml
[project]
name = "rag"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchainhub>=0.1.21",
    "pylint>=4.0.5",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "routes>=2.5.1",
    "uvicorn>=0.44.0",
]
